moved tests to ops package
