+++ dataset/src/test/scala/frameless/ops/RollupTests.scala
@@ -1,20 +1,20 @@
 package frameless
-package forward
+package ops
 
 import frameless.functions.aggregate._
 import org.scalacheck.Prop
 import org.scalacheck.Prop._
 
-class CubeTests extends TypedDatasetSuite {
+class RollupTests extends TypedDatasetSuite {
 
-  test("cube('a).agg(count())") {
+  test("rollup('a).agg(count())") {
     def prop[A: TypedEncoder : Ordering, Out: TypedEncoder : Numeric]
     (data: List[X1[A]])(implicit summable: CatalystSummable[A, Out]): Prop = {
       val dataset = TypedDataset.create(data)
       val A = dataset.col[A]('a)
 
-      val received = dataset.cube(A).agg(count()).collect().run().toVector.sortBy(_._2)
-      val expected = dataset.dataset.cube("a").count().collect().toVector
+      val received = dataset.rollup(A).agg(count()).collect().run().toVector.sortBy(_._2)
+      val expected = dataset.dataset.rollup("a").count().collect().toVector
         .map(row => (Option(row.getAs[A](0)), row.getAs[Long](1))).sortBy(_._2)
 
       received ?= expected
@@ -23,15 +23,15 @@ class CubeTests extends TypedDatasetSuite {
     check(forAll(prop[Int, Long] _))
   }
 
-  test("cube('a, 'b).agg(count())") {
+  test("rollup('a, 'b).agg(count())") {
     def prop[A: TypedEncoder : Ordering, B: TypedEncoder, Out: TypedEncoder : Numeric]
     (data: List[X2[A, B]])(implicit summable: CatalystSummable[B, Out]): Prop = {
       val dataset = TypedDataset.create(data)
       val A = dataset.col[A]('a)
       val B = dataset.col[B]('b)
 
-      val received = dataset.cube(A, B).agg(count()).collect().run().toVector.sortBy(_._3)
-      val expected = dataset.dataset.cube("a", "b").count().collect().toVector
+      val received = dataset.rollup(A, B).agg(count()).collect().run().toVector.sortBy(_._3)
+      val expected = dataset.dataset.rollup("a", "b").count().collect().toVector
         .map(row => (Option(row.getAs[A](0)), Option(row.getAs[B](1)), row.getAs[Long](2))).sortBy(_._3)
 
       received ?= expected
@@ -40,15 +40,15 @@ class CubeTests extends TypedDatasetSuite {
     check(forAll(prop[Int, Long, Long] _))
   }
 
-  test("cube('a).agg(sum('b)") {
+  test("rollup('a).agg(sum('b)") {
     def prop[A: TypedEncoder : Ordering, B: TypedEncoder, Out: TypedEncoder : Numeric]
     (data: List[X2[A, B]])(implicit summable: CatalystSummable[B, Out]): Prop = {
       val dataset = TypedDataset.create(data)
       val A = dataset.col[A]('a)
       val B = dataset.col[B]('b)
 
-      val received = dataset.cube(A).agg(sum(B)).collect().run().toVector.sortBy(_._2)
-      val expected = dataset.dataset.cube("a").sum("b").collect().toVector
+      val received = dataset.rollup(A).agg(sum(B)).collect().run().toVector.sortBy(_._2)
+      val expected = dataset.dataset.rollup("a").sum("b").collect().toVector
         .map(row => (Option(row.getAs[A](0)), row.getAs[Out](1))).sortBy(_._2)
 
       received ?= expected
@@ -57,13 +57,13 @@ class CubeTests extends TypedDatasetSuite {
     check(forAll(prop[Int, Long, Long] _))
   }
 
-  test("cube('a).mapGroups('a, sum('b))") {
+  test("rollup('a).mapGroups('a, sum('b))") {
     def prop[A: TypedEncoder : Ordering, B: TypedEncoder : Numeric]
     (data: List[X2[A, B]]): Prop = {
       val dataset = TypedDataset.create(data)
       val A = dataset.col[A]('a)
 
-      val received = dataset.cube(A)
+      val received = dataset.rollup(A)
         .deserialized.mapGroups { case (a, xs) => (a, xs.map(_.b).sum) }
         .collect().run().toVector.sortBy(_._1)
       val expected = data.groupBy(_.a).mapValues(_.map(_.b).sum).toVector.sortBy(_._1)
@@ -74,7 +74,7 @@ class CubeTests extends TypedDatasetSuite {
     check(forAll(prop[Int, Long] _))
   }
 
-  test("cube('a).agg(sum('b), sum('c)) to cube('a).agg(sum('a), sum('b), sum('a), sum('b), sum('a))") {
+  test("rollup('a).agg(sum('b), sum('c)) to rollup('a).agg(sum('a), sum('b), sum('a), sum('b), sum('a))") {
     def prop[
     A: TypedEncoder : Ordering,
     B: TypedEncoder,
@@ -92,38 +92,38 @@ class CubeTests extends TypedDatasetSuite {
       val C = dataset.col[C]('c)
 
       val framelessSumBC = dataset
-        .cube(A)
+        .rollup(A)
         .agg(sum(B), sum(C))
         .collect().run().toVector.sortBy(_._1)
 
-      val sparkSumBC = dataset.dataset.cube("a").sum("b", "c").collect().toVector
+      val sparkSumBC = dataset.dataset.rollup("a").sum("b", "c").collect().toVector
         .map(row => (Option(row.getAs[A](0)), row.getAs[OutB](1), row.getAs[OutC](2)))
         .sortBy(_._1)
 
       val framelessSumBCB = dataset
-        .cube(A)
+        .rollup(A)
         .agg(sum(B), sum(C), sum(B))
         .collect().run().toVector.sortBy(_._1)
 
-      val sparkSumBCB = dataset.dataset.cube("a").sum("b", "c", "b").collect().toVector
+      val sparkSumBCB = dataset.dataset.rollup("a").sum("b", "c", "b").collect().toVector
         .map(row => (Option(row.getAs[A](0)), row.getAs[OutB](1), row.getAs[OutC](2), row.getAs[OutB](3)))
         .sortBy(_._1)
 
       val framelessSumBCBC = dataset
-        .cube(A)
+        .rollup(A)
         .agg(sum(B), sum(C), sum(B), sum(C))
         .collect().run().toVector.sortBy(_._1)
 
-      val sparkSumBCBC = dataset.dataset.cube("a").sum("b", "c", "b", "c").collect().toVector
+      val sparkSumBCBC = dataset.dataset.rollup("a").sum("b", "c", "b", "c").collect().toVector
         .map(row => (Option(row.getAs[A](0)), row.getAs[OutB](1), row.getAs[OutC](2), row.getAs[OutB](3), row.getAs[OutC](4)))
         .sortBy(_._1)
 
       val framelessSumBCBCB = dataset
-        .cube(A)
+        .rollup(A)
         .agg(sum(B), sum(C), sum(B), sum(C), sum(B))
         .collect().run().toVector.sortBy(_._1)
 
-      val sparkSumBCBCB = dataset.dataset.cube("a").sum("b", "c", "b", "c", "b").collect().toVector
+      val sparkSumBCBCB = dataset.dataset.rollup("a").sum("b", "c", "b", "c", "b").collect().toVector
         .map(row => (Option(row.getAs[A](0)), row.getAs[OutB](1), row.getAs[OutC](2), row.getAs[OutB](3), row.getAs[OutC](4), row.getAs[OutB](5)))
         .sortBy(_._1)
 
@@ -136,7 +136,7 @@ class CubeTests extends TypedDatasetSuite {
     check(forAll(prop[String, Long, BigDecimal, Long, BigDecimal] _))
   }
 
-  test("cube('a, 'b).agg(sum('c), sum('d))") {
+  test("rollup('a, 'b).agg(sum('c), sum('d))") {
     def prop[
     A: TypedEncoder : Ordering,
     B: TypedEncoder : Ordering,
@@ -156,12 +156,12 @@ class CubeTests extends TypedDatasetSuite {
       val D = dataset.col[D]('d)
 
       val framelessSumByAB = dataset
-        .cube(A, B)
+        .rollup(A, B)
         .agg(sum(C), sum(D))
         .collect().run().toVector.sortBy(x => (x._1, x._2))
 
       val sparkSumByAB = dataset.dataset
-        .cube("a", "b").sum("c", "d").collect().toVector
+        .rollup("a", "b").sum("c", "d").collect().toVector
         .map(row => (Option(row.getAs[A](0)), Option(row.getAs[B](1)), row.getAs[OutC](2), row.getAs[OutD](3)))
         .sortBy(x => (x._1, x._2))
 
@@ -171,7 +171,7 @@ class CubeTests extends TypedDatasetSuite {
     check(forAll(prop[Byte, Int, Long, BigDecimal, Long, BigDecimal] _))
   }
 
-  test("cube('a, 'b).mapGroups('a, 'b, sum('c))") {
+  test("rollup('a, 'b).mapGroups('a, 'b, sum('c))") {
     def prop[
     A: TypedEncoder : Ordering,
     B: TypedEncoder : Ordering,
@@ -182,7 +182,7 @@ class CubeTests extends TypedDatasetSuite {
       val B = dataset.col[B]('b)
 
       val framelessSumByAB = dataset
-        .cube(A, B)
+        .rollup(A, B)
         .deserialized.mapGroups { case ((a, b), xs) => (a, b, xs.map(_.c).sum) }
         .collect().run().toVector.sortBy(x => (x._1, x._2))
 
@@ -196,7 +196,7 @@ class CubeTests extends TypedDatasetSuite {
     check(forAll(prop[Byte, Int, Long] _))
   }
 
-  test("cube('a).mapGroups(('a, toVector(('a, 'b))") {
+  test("rollup('a).mapGroups(('a, toVector(('a, 'b))") {
     def prop[
     A: TypedEncoder,
     B: TypedEncoder
@@ -205,7 +205,7 @@ class CubeTests extends TypedDatasetSuite {
       val A = dataset.col[A]('a)
 
       val datasetGrouped = dataset
-        .cube(A)
+        .rollup(A)
         .deserialized.mapGroups((a, xs) => (a, xs.toVector))
         .collect().run().toMap
 
@@ -219,7 +219,7 @@ class CubeTests extends TypedDatasetSuite {
     check(forAll(prop[X1[Option[Short]], Short] _))
   }
 
-  test("cube('a).flatMapGroups(('a, toVector(('a, 'b))") {
+  test("rollup('a).flatMapGroups(('a, toVector(('a, 'b))") {
     def prop[
     A: TypedEncoder : Ordering,
     B: TypedEncoder : Ordering
@@ -228,7 +228,7 @@ class CubeTests extends TypedDatasetSuite {
       val A = dataset.col[A]('a)
 
       val datasetGrouped = dataset
-        .cube(A)
+        .rollup(A)
         .deserialized.flatMapGroups((a, xs) => xs.map(x => (a, x)))
         .collect().run()
         .sorted
@@ -246,7 +246,7 @@ class CubeTests extends TypedDatasetSuite {
     check(forAll(prop[X1[Option[Short]], Short] _))
   }
 
-  test("cube('a, 'b).flatMapGroups((('a,'b) toVector((('a,'b), 'c))") {
+  test("rollup('a, 'b).flatMapGroups((('a,'b) toVector((('a,'b), 'c))") {
     def prop[
     A: TypedEncoder : Ordering,
     B: TypedEncoder : Ordering,
@@ -257,13 +257,13 @@ class CubeTests extends TypedDatasetSuite {
       val cB = dataset.col[B]('b)
 
       val datasetGrouped = dataset
-        .cube(cA, cB)
+        .rollup(cA, cB)
         .deserialized.flatMapGroups((a, xs) => xs.map(x => (a, x)))
         .collect().run()
         .sorted
 
       val dataGrouped = data
-        .groupBy(t => (t.a,t.b)).toSeq
+        .groupBy(t => (t.a, t.b)).toSeq
         .flatMap { case (a, xs) => xs.map(x => (a, x)) }
         .sorted
 
@@ -275,14 +275,14 @@ class CubeTests extends TypedDatasetSuite {
     check(forAll(prop[X1[Option[Short]], Short, Byte] _))
   }
 
-  test("cubeMany('a).agg(sum('b))") {
+  test("rollupMany('a).agg(sum('b))") {
     def prop[A: TypedEncoder : Ordering, Out: TypedEncoder : Numeric]
     (data: List[X1[A]])(implicit summable: CatalystSummable[A, Out]): Prop = {
       val dataset = TypedDataset.create(data)
       val A = dataset.col[A]('a)
 
-      val received = dataset.cubeMany(A).agg(count()).collect().run().toVector.sortBy(_._2)
-      val expected = dataset.dataset.cube("a").count().collect().toVector
+      val received = dataset.rollupMany(A).agg(count()).collect().run().toVector.sortBy(_._2)
+      val expected = dataset.dataset.rollup("a").count().collect().toVector
         .map(row => (Option(row.getAs[A](0)), row.getAs[Long](1))).sortBy(_._2)
 
       received ?= expected
